refactor: implement authentication and authorization middleware for user routes
