+++ apps/server/src/middlewares/auth.middleware.ts
@@ -1,0 +1,94 @@
+import { prisma } from "@repo/database";
+import { UserPublicInfo } from "@repo/domain";
+import { NextFunction, Request, RequestHandler, Response } from "express";
+import jwt from "jsonwebtoken";
+import AppError from "../utils/appError.js";
+import catchAsync from "../utils/catchAsync.js";
+import { env } from "../utils/env.js";
+
+/**
+ * Middleware to authenticate users via JWT token.
+ * Checks for token in Authorization header (Bearer) or cookies.
+ * Validates token, checks if user exists, and verifies password hasn't changed since token issuance.
+ * Attaches user to req.user on success.
+ */
+export const authenticate: RequestHandler = catchAsync(
+  async (req, _res, next) => {
+    // * 1) Getting token and check if it's there
+    const { authorization } = req.headers;
+    let token;
+    if (authorization?.startsWith("Bearer")) {
+      token = authorization.replace("Bearer ", "");
+    } else if (req.cookies?.jwt) {
+      token = req.cookies.jwt;
+    }
+    if (!token) {
+      return next(
+        new AppError("You ar not logged in! Please log in to again access", 401)
+      );
+    }
+
+    //* 2) Validate Token
+    const decoded = jwt.verify(token, env.JWT_SECRET);
+    if (!decoded || typeof decoded === "string") {
+      return next(new AppError("Invalid token", 401));
+    }
+    // //* 3) check if user still exists
+    const currentUser = await prisma.user.findUnique({
+      where: { id: decoded.id },
+    });
+    if (!currentUser) {
+      return next(
+        new AppError(
+          "The user belonging to this token does no longer exists",
+          401
+        )
+      );
+    }
+    // //* 4) check if user changed password after the token was issued
+    const hasPasswordChanged = await prisma.user.isPasswordChangedAfter(
+      decoded.iat!,
+      currentUser.passwordChangedAt
+    );
+    if (hasPasswordChanged) {
+      return next(
+        new AppError(
+          "User recently changed password! Please log in to again",
+          401
+        )
+      );
+    }
+
+    req.user = currentUser;
+
+    return next();
+  }
+);
+
+/**
+ * Middleware factory to authorize users based on roles.
+ * Must be used after authenticate middleware.
+ *
+ * @param roles - Allowed roles for the route
+ * @returns Middleware that checks if req.user.role matches any of the allowed roles
+ *
+ * @example
+ * router.delete('/admin-only', authenticate, authorize('ADMIN'), handler);
+ */
+export const authorize = (...roles: UserPublicInfo["role"][]) => {
+  return (req: Request, _res: Response, next: NextFunction) => {
+    if (!req.user) {
+      return next(
+        new AppError("Authentication required before authorization", 401)
+      );
+    }
+
+    if (!roles.includes(req.user.role)) {
+      return next(
+        new AppError("You do not have permission to perform this action", 403)
+      );
+    }
+
+    next();
+  };
+};
