+++ apps/server/src/routes/config.route.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import * as configController from "../controllers/config.controller.js";
-import * as authController from "../controllers/auth.controller.js";
+import { authenticate, authorize } from "../middlewares/auth.middleware.js";
 import { validateSchema } from "../middlewares/validation.middleware.js";
 import { CreateConfigSchema } from "@repo/domain";
 
@@ -9,10 +9,7 @@ const configRouter: express.Router = express.Router();
 configRouter.get("/", configController.getConfig);
 
 // * ADMIN ROUTES (restricted to admin roles)
-configRouter.use(
-  authController.authenticate,
-  authController.checkAuthorization("ADMIN")
-);
+configRouter.use(authenticate, authorize("ADMIN"));
 
 configRouter.post(
   "/",
